server/requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-money==3.5.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
py-moneyed==3.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
